Double Jump Added
Added a double jump, changed the jump mechanic from using add force vector to new vectors. The set-up for the double jump is a little weird with how brackey's code is set up, should refactor so both jumps are in the same place. Currently the first jump is in the move function, while the second jump is in the movementscript update.