Pushing changes to Animator & Movement script.
Fixed Animator so that wall climbing/clinging animation is used while player is using respective power.  Altered Movement Script so that animator bool for "Jump" is set false, allowing climbing/clinging animation to be seen.  Animation is currently stuck looping on first frame of animation.  Serviceable but not ideal.
Updated project details, changed version to 0.1.0 and product name to "Fragment"